Remove items reward
* added "remove items from inventory" reward
* reworked Quest Object API to use a POJO instead of parameters
* added colors to pool editor messages, using a new "prefix" system for
Lang enum
* reworked time editor for quest pools
* fixed an issue with the minecraftTranslationsFile config option

+++ core/src/main/java/fr/skytasul/quests/options/OptionTimer.java
@@ -54,8 +54,8 @@ public class OptionTimer extends QuestOption<Integer> {
 	@Override
 	public void click(FinishGUI gui, Player p, ItemStack item, int slot, ClickType click) {
 		Lang.TIMER.send(p);
-		new TextEditor<>(p, () -> gui.reopen(p), (obj) -> {
-			setValue(obj);
+		new TextEditor<>(p, () -> gui.reopen(p), obj -> {
+			setValue(obj.intValue());
 			ItemUtils.lore(item, getLore());
 			gui.reopen(p);
 		}, () -> {
